+++ src/def/IAccountJailInfo.d.ts
@@ -1,4 +1,12 @@
 interface IAccountJailInfo {
+    // Contain all user attempts
+    attempts : ILocalProtocolAccountAttempt[];
+
     // Contains the date with lock action expire
     endLock : number;
+}
+
+interface IAccountJailInfoPublic extends IAccountJailInfo {
+    // Account id locked
+    accountID : string;
 }