+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## Change Log
 
-### 
-- Local Protocol :: Add lockAccount/unlockAccount method
+### v0.1.0
+- Local Protocol :: Protocol is now implemented correctly
+- Local Protocol :: Add getAccounts/getUsers/getBannedUsers/getLockedAccounts/lockAccount/unlockAccount method
 - Local Protocol :: Lock account if too many connection attempts have been sent on
 - Local Protocol :: Add more functional test
-
+- Trapjs :: Review unbanUser/unlockAccount methods to use string or array
 
 ### v0.0.5 (2015/05/10)
 - Implement Local protocol